backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
gunicorn==21.2.0
pdfplumber==0.10.2
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
nltk==3.8.1
spacy==3.7.2
pydantic==2.4.2
whitenoise==6.5.0 